B023 Minor changes per checklist Eliminated eslint-ignore'd issues, moved rate-limiter configurations to a file, renamed routes to index.
