+++ middleware/rate-limiter.js
@@ -1,0 +1,6 @@
+const rateLimit = require('express-rate-limit');
+
+module.exports = rateLimit({
+  windowMs: 15 * 60 * 1000,
+  max: 100,
+});
